mailbox: kvx: Fix mbox value width
Summary:
The mailbox can write a 64bit value, hence read 64bit from data

ref T20632

Test Plan: for optee

Reviewers: O51 Linux Coolidge!

Subscribers: #linux_coolidge_cc

Maniphest Tasks: T20632

Differential Revision: https://phab.kalray.eu/D28080

+++ drivers/mailbox/kvx-mailbox.c
@@ -51,7 +51,7 @@ static irqreturn_t kvx_mbox_isr(int irq, void *p)
 static int kvx_mbox_send_data(struct mbox_chan *chan, void *data)
 {
 	struct kvx_mbox_priv *mbox = to_kvx_mbox_priv(chan->mbox);
-	uint64_t mbox_val = *(uint32_t *) data;
+	uint64_t mbox_val = *(uint64_t *) data;
 
 	if (mbox->dir != MBOX_DIR_TX)
 		return -EINVAL;
